Remove redundant cast to int

+++ graalpython/com.oracle.graal.python/src/com/oracle/graal/python/builtins/objects/cext/hpy/llvm/GraalHPyLLVMNodes.java
@@ -306,7 +306,7 @@ abstract class GraalHPyLLVMNodes {
                 throw CompilerDirectives.shouldNotReachHere("returned pointer object must have array type");
             }
 
-            Object[] elements = new Object[(int) n];
+            Object[] elements = new Object[n];
             try {
                 for (int i = 0; i < elements.length; i++) {
                     /*
